fix(Catalog): 目录节点拖拽排序支持浮点 sort_order，修复 422 Unprocessable Entity 错误; (#865)
前端采用分数索引（Fractional Indexing）策略计算拖拽目标位置，取相邻节点 sort_order
的中点（如 812.5）。但后端全栈（Pydantic schema → Service → DAO → DB）均将
sort_order/position 声明为 int/Integer，导致 Pydantic 在 API 边界拒绝浮点值。

变更：
- Pydantic schemas: sort_order 类型从 int 改为 float（Create/Update/Response）
- Dataclass 类型: CatalogNodeRecord/CatalogTreeItem 的 sort_order 改为 float
- Service/DAO 层: 方法参数类型从 int 改为 float
- DB Model: position 列从 Integer 改为 Float
- 新增 Alembic migration 0061: ALTER COLUMN position TYPE DOUBLE PRECISION

+++ apps/negentropy/src/negentropy/knowledge/lifecycle_schemas.py
@@ -79,7 +79,7 @@ class CatalogNodeCreateRequest(BaseModel):
     parent_id: UUID | None = None
     node_type: str = "folder"
     description: str | None = None
-    sort_order: int = 0
+    sort_order: float = 0
     config: dict[str, Any] = Field(default_factory=dict)
 
     @field_validator("node_type")
@@ -106,7 +106,7 @@ class CatalogNodeUpdateRequest(BaseModel):
     parent_id: UUID | None = None
     node_type: str | None = None
     description: str | None = None
-    sort_order: int | None = None
+    sort_order: float | None = None
     config: dict[str, Any] | None = None
 
 
@@ -122,7 +122,7 @@ class CatalogNodeResponse(BaseModel):
     slug: str
     node_type: str
     description: str | None = None
-    sort_order: int
+    sort_order: float
     config: dict[str, Any] = Field(default_factory=dict)
     # DOCUMENT_REF 叶子节点携带的文档 ID（FOLDER / CATEGORY / COLLECTION 为 None）
     document_id: UUID | None = None
